RAG/requirements.txt
google-genai
pandas
numpy
dotenv
os
